% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addUtilsFunctions.R
\name{addLintrConfig}
\alias{addLintrConfig}
\title{Copy lintr configuration to package}
\usage{
addLintrConfig(outFile = ".lintr")
}
\arguments{
\item{outFile}{Character scalar, giving the path to which the script 
should be copied. The path is relative to the root of the active 
project. If a file with this name already exists, the function will 
ask for confirmation before overwriting it. \code{outFile} will also 
be added to the \code{.Rbuildignore} and \code{.gitignore} files.}
}
\description{
This function copies a .lintr configuration file to a new script in a 
specified location.
}
\author{
Charlotte Soneson
}
