% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addUtilsFunctions.R
\name{addUtilsFunctions}
\alias{addUtilsFunctions}
\title{Copy utility functions to package}
\usage{
addUtilsFunctions(outFile = "R/utils.R", copyTests = TRUE)
}
\arguments{
\item{outFile}{Character scalar, giving the path to which the script 
should be copied. The path is relative to the root of the active 
project. If a file with this name already exists, the function will 
ask for confirmation before overwriting it.}

\item{copyTests}{Logical scalar, defining whether to copy unit tests for 
the utility functions to \code{tests/testthat/test-<name>.R}, 
where \code{<name>} is the base name of \code{outFile}. If the 
target package is not yet set up to use \code{testthat}, the function 
will also run \code{usethis::use_testthat()} to generate the required 
folder structure and add \code{testthat} to the list of suggested 
package in the DESCRIPTION file.}
}
\description{
This function copies handy utility functions to a new script in a 
specified location. Currently, the script contains the following utility 
functions:
\itemize{
\item \code{.assertScalar()} - convenience function to check the validity 
    of scalar variables.
\item \code{.assertVector()} - convenience function to check the validity 
    of vector variables. 
}
}
\author{
Charlotte Soneson
}
