% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calcAndCountDist}
\alias{calcAndCountDist}
\title{Count frequency of differences between values in integer vectors.}
\usage{
calcAndCountDist(query, reference, cnt)
}
\arguments{
\item{query}{first \code{integer} vector.}

\item{reference}{second \code{integer} vector. Distances are calculated from each element
in \code{query} to each greater element in \code{reference}.}

\item{cnt}{\code{NumericVector} to store the result in. The length of \code{cnt} defines
the maximal distance that will be included in the analysis, and new counts
will be added to the values of \code{cnt}.}
}
\value{
\code{numeric} vector \code{cnt}, where \code{cnt[d]} correspond to
  the number of observed distances \code{d}.
}
\description{
Given two ascendingly sorted integer vectors \code{query} and \code{reference}, calculate
  and count the differences between their elements that are greater than zero
  and less than \code{maxd}. The number of observed distances \code{d} are reported in \code{cnt[d]},
  and \code{maxd} corresponds to the \code{length(cnt)}. The function is
  called by \code{\link{calcPhasogram}}, which provides a higher level, more
  conventient interface.
}
\author{
Michael Stadler
}
