% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colByValue.R
\name{colByValue}
\alias{colByValue}
\title{Map numerical values to colors.}
\usage{
colByValue(x, rng = range(x, na.rm = TRUE), col = c("#5E4FA2",
  "#3288BD", "#66C2A5", "#ABDDA4", "#E6F598", "#FFFFBF", "#FEE08B",
  "#FDAE61", "#F46D43", "#D53E4F", "#9E0142"), NA.col = "lightgray",
  alpha = 255)
}
\arguments{
\item{x}{\code{numeric} vector with values to be mapped to colors.}

\item{rng}{\code{numeric(2)} giving the range of values to be mapped to colors.
By default, this will be the range of finite values in \code{x}.}

\item{col}{vector with R colors defining the palette (must be a valid argument
to \code{\link[grDevices]{col2rgb}}.}

\item{NA.col}{Single R color to use for \code{NA} values in \code{x}.}

\item{alpha}{\code{numeric(1)} between 0 and 255, giving the alpha channel
value for the colors (0 = fully transparent, 255 = fully opaque,
ignored if \code{col} already contain alpha values).}
}
\value{
A \code{character} vector of the same length of \code{x} with
    R colors in hexadecimal string-encoded RGB format.
}
\description{
\code{colByValue} takes a numerical vector and maps each value
    to an R color string.
}
\details{
The values in \code{[rng[1], rng[2]]} will be linearly mapped to the
    color palette defined by \code{col}. Any values in \code{x} less (greater)
    than \code{rng[1]} (\code{rng[2]}) will be assigned the same color as
    \code{rng[1]} (\code{rng[2]}).
}
\examples{
x <- rnorm(1000)
y <- rnorm(1000)
cols <- colByValue(x + y)
plot(x, y, pch = 20, col = cols, main = "default")

}
\seealso{
\code{\link[grDevices]{colorRamp}} and \code{\link[grDevices]{rgb}}
    for the functions called by \code{colByValue}.
}
\author{
Michael Stadler
}
