% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.assertPackagesAvailable}
\alias{.assertPackagesAvailable}
\title{Make sure that packages are available}
\usage{
.assertPackagesAvailable(pkgs, bioc = TRUE)
}
\arguments{
\item{pkgs}{Character vector with package names.}

\item{bioc}{Logical scalar. If \code{TRUE}, suggest to install the missing
package(s) using \code{BiocManager} instead of \code{install.packages}.}
}
\description{
Try loading the namespaces of the packages given in \code{pkgs}, and
throw an exception with an informative error message if that is not
the case.
}
\author{
Michael Stadler
}
