% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valueToColor.R
\name{.isValidColor}
\alias{.isValidColor}
\title{Check for valid color specification.}
\usage{
.isValidColor(x)
}
\arguments{
\item{x}{\code{numeric} or \code{character} vector with color specifications.}
}
\value{
A \code{logical} vector of the same length as \code{x}.
}
\description{
Check if the elements of a \code{numeric} or \code{character} vector
    are valid colors understood by R.
}
\details{
This is an internal function, used for example by \code{valueToColor}.
    The elements of \code{x} are check for valid color specifications by
    passing them to \code{col2rgb} and checking for an exception.
}
\examples{
swissknife:::.isValidColor(c("black", "blackk"))
swissknife:::.isValidColor(c(-1, 1))

}
\author{
Michael Stadler
}
\keyword{internal}
