% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phasograms.R
\name{estimateNRL}
\alias{estimateNRL}
\title{Estimate the nucleosome repeat length (NRL) from a phasogram.}
\usage{
estimateNRL(
  x,
  mind = 140L,
  usePeaks = 1:8,
  span1 = 100/length(x),
  span2 = 1500/length(x)
)
}
\arguments{
\item{x}{\code{numeric} vector giving the counts of alignment distances
(typically the output of \code{\link{calcPhasogram}}.}

\item{mind}{\code{integer(1)} specifying the minimal distance to be used for
NRL estimation. The default value (140) ignores any distance too short to
span at least a single nucleosome.}

\item{usePeaks}{\code{integer} vector selecting the modes (peaks) in the
phasogram used in NRL estimation.}

\item{span1}{\code{numeric(1)} giving the smoothing parameter for de-trending
loess fit (high pass filter).}

\item{span2}{\code{numeric(1)} giving the smoothing parameter for de-noising
loess fit (low pass filter).}
}
\value{
A \code{list} with elements: \describe{
  \item{nrl}{the estimated nucleosome repeat length}
  \item{nrl.CI95}{the 95\% confidence interval}
  \item{xs}{smoothed (de-trended) phasogram}
  \item{loessfit}{the de-noising fit to the de-trended phasogram}
  \item{lmfit}{the linear fit to the phasogram peaks}
  \item{peaks}{the peak locations}
  \item{mind}{minimal distance included in the fit}
  \item{span1}{smoothing parameter for de-trending loess fit}
  \item{span2}{smoothing parameter for de-noising loess fit}
  \item{usePeaks}{the peaks used in the fit}}
}
\description{
Estimate the nucleosome repeat length (NRL) from the frequencies
  of same-strand alignment distances (phasogram), e.g. generated by
  \code{\link{calcPhasogram}}. The NRL is obtained from the slope of a linear
  fit to the modes in the phasogram.
}
\examples{
  # see the help for calcPhasogram() for a full example

}
\seealso{
\code{\link{calcPhasogram}} to calculate the phasogram from
  alignments, \code{\link{plotPhasogram}} to visualize an annotated phasogram
}
\author{
Michael Stadler
}
