% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleData.R
\name{exampleData}
\alias{exampleData}
\title{Access example data}
\usage{
exampleData(name = "list", envir = globalenv(), verbose = TRUE)
}
\arguments{
\item{name}{An optional character scalar specifying the data set(s) to
be made available. The special name \code{"list"} (default) is used to
print a data frame of available data sets with descriptions. The special
name \code{"new"} will select the newest data set(s) available.}

\item{envir}{specifies the environment in which the data should be made
available. By default, \code{envir = globalenv()}, which creates the
example data objects in the user workspace. Possible alternative
environment are for example \code{parent.frame()}, which is the
environment in which \code{exampleData()} was called.}

\item{verbose}{A logical scalar. If \code{TRUE}, report what is being
selected and made available.}
}
\value{
A \code{data.frame} (invisibly) with one row for each dataset
    that was made available in the global environment.
}
\description{
Make example data available, typically for use in teaching.
}
\examples{
exampleData()
exampleData("mycars")

}
\author{
Michael Stadler
}
