% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selVarGenes.R
\name{getDistMat}
\alias{getDistMat}
\title{Get Distance}
\usage{
getDistMat(
  loessModel = NULL,
  x = NULL,
  y = NULL,
  x_curve = NULL,
  method = "euclidean",
  ...
)
}
\arguments{
\item{loessModel}{the loess fit to all points.}

\item{x}{the x values of all points.}

\item{y}{the y values of all points.}

\item{x_curve}{the x values of the points on the loess curve.}

\item{method}{distance measure to be used.}

\item{...}{additional parameters for the \code{dist.matrix} function.}
}
\value{
An object of class \code{matrix} and \code{dist.matrix}, containing the
   distances between all points to the specified points on the loess curve.
}
\description{
This function gets the euclidean (by default) distances between all points in a 2-D coordinate 
  system (x,y) to a group of points that fall on a fitted loess curve. It uses the \code{dist.matrix} 
  function from the \code{wordspace} package.It is used in the \code{selVarGenes} function.
}
\examples{
  library(stats)
  
  # create data
  x <- 1:1000
  y <- sort(rnorm(n = 1000, mean = 0, sd = 5)) + runif(n = 1000, min = 0, max = 10)
  lo <- loess(y ~ x)
  x_curve <- seq(range(x)[1], range(x)[2], length.out = 100)
  
  # plot
  plot(x, y, pch = 16, col = "#BEBEBE90")
  lines(x_curve, predict(lo, newdata = x_curve), col = "red")
  
  # get euclidean distance of all points to the curve (red)
  euclMat <- .getDistMat(loessModel = lo, x = x, y = y, x_curve = x_curve)
  euclMat[1:6, 1:6]

}
\author{
Dania Machlab
}
\keyword{INTERNAL}
