% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGenomicTiles.R
\name{getGenomicTiles}
\alias{getGenomicTiles}
\title{Get regions tiling a genome.}
\usage{
getGenomicTiles(
  genome,
  tileWidth,
  hasOverlap = list(),
  fracOverlap = list(),
  numOverlap = list(),
  nearest = list(),
  addSeqComp = TRUE
)
}
\arguments{
\item{genome}{The genome to work on. Either a \code{\link[BSgenome]{BSgenome}}
object, a \code{character} scalar with the name of an installed \code{\link[BSgenome]{BSgenome}}
or with a file path and name pointing to a fasta file with the genome sequence,
or a named \code{numeric} vector giving the names and lengths of chromosomes.}

\item{tileWidth}{\code{numeric} scalar with the tile length.}

\item{hasOverlap, fracOverlap, numOverlap, nearest}{Named \code{list}s with
\code{\link[GenomicRanges]{GRanges}} or \code{\link[GenomicRanges]{GRangesList}}
object(s) used to annotate genomic tiles. See \code{\link{annotateRegions}}
for details.}

\item{addSeqComp}{\code{logical} scalar. If \code{TRUE} and primary sequence
can be obtained from \code{genome}, also add sequence composition features
for each tile to the annotations. Currently, the following features are
included: percent of G+C bases ("percGC"), CpG observed-over-expected ratio
("CpGoe").}
}
\value{
A \code{\link[GenomicRanges]{GRanges}} object with genome tiling regions.
    Optional tile annotations are contained in its metadata columns (\code{mcols}).
}
\description{
Get sequential, potentially annotated regions of a fixed lengths
   (tiles) along chromosomes of a genome.
}
\details{
The last tile in each chromosome is dropped if it would be shorter
    than \code{tileWidth}. Generated tiles are unstranded (\code{*}) and
    therefore overlaps or searching for nearest neighbors are ignoring
    strands of annotations (\code{ignore.strand=TRUE}).
}
\examples{
library(GenomicRanges)

tss <- GRanges("chr1", IRanges(c(1, 10, 30), width = 1,
                               names = paste0("t", 1:3)))
blacklist <- GRanges("chr1", IRanges(20, width = 5))
getGenomicTiles(c(chr1 = 45, chr2 = 12), tileWidth = 10,
                hasOverlap = list(Blacklist = blacklist),
                fracOverlap = list(Blacklist = blacklist),
                numOverlap = list(TSS = tss),
                nearest = list(TSS = tss))

}
\seealso{
\code{\link[GenomicRanges]{tileGenome}} and \code{\link{annotateRegions}}
    used by \code{getGenomicTiles} internally.
}
\author{
Michael Stadler
}
