% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInsertSizeDistFromBam.R
\name{getInsertSizeDistFromBam}
\alias{getInsertSizeDistFromBam}
\title{Tabulate insert sizes from paired-end alignments in bam files.}
\usage{
getInsertSizeDistFromBam(
  fname,
  regions = NULL,
  nmax = NA_integer_,
  isizemax = 800
)
}
\arguments{
\item{fname}{\code{character} vector with one or several bam files. If
multiple files are given, insert sizes from all will be pooled and
tabulated together.}

\item{regions}{\code{GRanges} object. Only alignments falling into these
regions will be used. If \code{NULL} (the default), all alignments are
used.}

\item{nmax}{\code{numeric(1)} specifying the maximal number of alignments to
read. If \code{NA} (the default), the alignments in \code{regions} (if
\code{regions} are not \code{NULL}) or in the bam file will be used.}

\item{isizemax}{\code{numeric(1)} specifying the maximal insert size to
report. Larger insert sizes will be set to \code{isizemax} with a warning.}
}
\value{
\code{integer} vector with the number of insert sizes. The element at
  position \code{i} gives the observed number of alignment pairs with an
  insert size of \code{i}. The number of insert sizes greater than
  \code{isizemax} that were set to \code{isizemax} are reported in the
  attribute \code{"ncapped"}.
}
\description{
Read and tabulate the insert sizes from paired-end alignments
  contained in one or several bam files. By default, all properly aligned
  read pairs are included. Optionally, alignments can be restricted to
  those in a specific genomic region region (\code{regions} argument) or
  the number of alignments read can be limited (\code{nmax} argument).
}
\examples{
if (requireNamespace("Rsamtools", quietly = TRUE)) {
    bamf <- system.file("extdata", "getInsertSizeDistFromBam", "atac_mm10.bam",
                        package = "swissknife")
    isize <- getInsertSizeDistFromBam(bamf)
    attr(isize, "ncapped")
    plot(isize, type = "l",
         xlab = "Insert size (bp)", ylab = "Number of fragments")
}

}
\seealso{
\code{\link[Rsamtools]{scanBam}} used to read alignments.
}
\author{
Michael Stadler
}
