% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMappableRegions.R
\name{getMappableRegions}
\alias{getMappableRegions}
\title{Get mappable regions of a genome.}
\usage{
getMappableRegions(
  genome,
  genomeIndex,
  kmerLength = 50,
  maxHits = 1,
  Ncpu = 2,
  quiet = TRUE
)
}
\arguments{
\item{genome}{The genome sequence to work on. Either a \code{\link[BSgenome]{BSgenome}}
object, a \code{character} scalar with the name of an installed \code{\link[BSgenome]{BSgenome}}
or with a file path and name pointing to a fasta file with the genome sequence.}

\item{genomeIndex}{\code{character} scalar with the path to the bowtie index
and prefix to align against, in the form \code{</path/to/index>/<prefix>},
or the name of an installed \code{Rbowtie} index package created by the
\pkg{QuasR} package for an installed \code{BSgenome} package.}

\item{kmerLength}{\code{numeric} scalar specifying the k-mer length (width of
overlapping windows in \code{genome}), usually set to the typical read
length for which to get the mappable regions.}

\item{maxHits}{\code{numeric} scalar specifying the maximum number of hits
(matches) of a k-mer in the \code{genome} to be considered mappable.}

\item{Ncpu}{\code{numeric} scalar specifying the number of CPU threads to use
for alignments.}

\item{quiet}{\code{logical} scalar indicating if progress information should
be printed on the console.}
}
\value{
A \code{\link[GenomicRanges]{GRanges}} object with mappable regions.
    All plus-strand sequences in \code{genome} of length \code{kmerLength}
    with their start (leftmost) position overlapping the \code{GRanges} object
    do not generate more than \code{maxHits} hits when aligned to the genome.
}
\description{
Given a k-mer length and the maximum number of allowed hits per
    k-mer, find all mappable regions in a genome.
}
\details{
Sequences of all overlapping windows are extracted from the genome
    and aligned to the provided genome index using \code{\link[Rbowtie]{bowtie}}
    with parameters \code{-f -v 0 -a -B 1 -m maxHits}. If no more
    than \code{maxHits} hits are found, the window is defined mappable.
}
\examples{
library(Rbowtie)

genomefile <- system.file("extdata", "getMappableRegions", "hg19sub.fa", package = "swissknife")
indexdir <- tempfile()
indexpre <- "index"
indexname <- file.path(indexdir, indexpre)
idx <- bowtie_build(genomefile, indexdir)

mapgr <- getMappableRegions(genomefile, indexname, 50, quiet = FALSE)
mapgr

}
\seealso{
\code{\link[Rbowtie]{bowtie}} in package \pkg{Rbowtie} used by
    \code{getMappableRegions} to align reads to the genome;
    \code{\link[Rbowtie]{bowtie_build}} in package \pkg{Rbowtie} for
    indexing a genome.
}
\author{
Michael Stadler
}
