% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_labeling.R
\name{normGenesetExpression}
\alias{normGenesetExpression}
\title{Calculate normalized expression of a gene set}
\usage{
normGenesetExpression(
  sce,
  genes,
  expr_values = "logcounts",
  R = 200,
  BPPARAM = SerialParam()
)
}
\arguments{
\item{sce}{\code{SingleCellExperiment} object.}

\item{genes}{\code{character} vector with the genes in the set. Must be a
subset of \code{rownames(sce)}.}

\item{expr_values}{Integer scalar or string indicating which assay of
\code{sce} contains the expression values.}

\item{R}{Integer scalar giving the number of random gene sets to sample
for normalization.}

\item{BPPARAM}{An optional \code{\link[BiocParallel]{BiocParallelParam}}
instance determining the parallel back-end to be used during evaluation.}
}
\value{
A \code{numeric} vector with normalized gene set scores for each
  cell in \code{sce}.
}
\description{
Calculate normalized expression for a set of genes in each cell
  from a \code{SingleCellExperiment}, using random sets of similarly
  expressed genes as background to account for cell quality and
  sequencing depth.
}
\author{
Michael Stadler
}
