% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsePkgVersions.R
\name{parsePkgVersions}
\alias{parsePkgVersions}
\title{Parse R and R package versions from session informations}
\usage{
parsePkgVersions(infiles)
}
\arguments{
\item{infiles}{Character vector with text files (extension must be either
\code{.Rout} or \code{.md}), containing session information to parse out.}
}
\value{
A \code{list} of \code{list}s with one element in the outer list for
  each R version, contianing an innter list with elements \code{files} and
  \code{packages}.
}
\description{
The function parses the R version and R package versions from
  session information (created by \code{sessionInfo()}, tested with R 3.6) in
  files provided in \code{infiles}. Two types of files are currently
  supported: \itemize{ \item{Rout: }{Files containing R console output
  (created by \code{R CMD BATCH scrip.R output.Rout}} \item{md: }{Files
  containing markdown output created by \code{rmarkdown::render('input.Rmd',
  clean = FALSE)}, which will keep the intermediate \code{.md} file.}}
}
\examples{
f <- list.files(system.file("extdata", "parsePkgVersions",
                            package = "swissknife"),
                full.names = TRUE)
parsePkgVersions(f)

}
\author{
Michael Stadler
}
