% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGeneRegion.R
\name{plotGeneRegion}
\alias{plotGeneRegion}
\title{Plot gene region}
\usage{
plotGeneRegion(
  gtf = "",
  granges = NULL,
  chr = "",
  start = NA_real_,
  end = NA_real_,
  showgene = "",
  bigwigFiles = "",
  bigwigCond = "",
  geneTrackTitle = "Genes"
)
}
\arguments{
\item{gtf}{Character scalar, path to gtf file (tested with Ensembl/Gencode
files)}

\item{granges}{GRanges object, typically generated from a GTF file using the
\code{prepareGTF} function}

\item{chr}{Character scalar, name of chromosome to show}

\item{start, end}{Numeric scalars, start and end position of the region to
show}

\item{showgene}{Character scalar, the gene ID/name to display}

\item{bigwigFiles}{Named character vector, paths to bigwig files}

\item{bigwigCond}{Named character vector, the grouping of the bigwig files
(used for coloring of the coverage tracks)}

\item{geneTrackTitle}{Character scalar, name of the gene track}
}
\description{
Visualize the gene model for a gene of interest, or for all genes in a
provided region. Also show one or more coverage tracks based on bigwig
file(s).
}
\examples{
gtffile <- system.file("extdata/plotGeneRegion/mm10_ensembl98.gtf", 
                       package = "swissknife")
plotGeneRegion(gtf = gtffile, 
               showgene = "Tnfaip3")
               
bwf <- system.file("extdata/plotGeneRegion/mnase_mm10.bw", package = "swissknife")
names(bwf) <- "bwf1"
plotGeneRegion(gtf = gtffile, 
               bigwigFiles = bwf,
               chr = "chr10", start = 20000000, end = 20005000)
               
bwf2 <- c(bwf, bwf)
names(bwf2) <- c("bwf1", "bwf2")
bwc2 <- c("c1", "c2")
names(bwc2) <- names(bwf2)
plotGeneRegion(gtf = gtffile, bigwigFiles = bwf2, bigwigCond = bwc2, 
               showgene = "Map3k5")

}
\author{
Charlotte Soneson
}
