% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phasograms.R
\name{plotPhasogram}
\alias{plotPhasogram}
\title{Plot annotated phasogram.}
\usage{
plotPhasogram(x, hide = TRUE, xlim = NULL, verbosePlot = FALSE, ...)
}
\arguments{
\item{x}{\code{numeric} vector giving the counts of alignment distances (typically
the output of \code{\link{calcPhasogram}}.}

\item{hide}{If \code{TRUE} (the default), hide phasogram counts not used in the
NRL estimate (\code{mind} parameter from \code{\link{estimateNRL}}).}

\item{xlim}{\code{numeric(2)} with the x-axis (phase) limits in the first two plots
(see Details). if \code{NULL} (the default), the full range defined by
\code{x} and \code{hide} will be used.}

\item{verbosePlot}{If \code{TRUE}, create three plots instead of just a single plot
(see Details).}

\item{...}{Additional arguments passed to \code{\link{estimateNRL}} to control NRL estimation.}
}
\value{
The return value from the call to \code{\link{estimateNRL}} (invisibly).
}
\description{
Plot phasogram and annotate it with estimated nucleosome repeat
    length (NRL).
}
\details{
The function will visualize an annotated phasogram. For \code{verbosePlot=FALSE}
    (the default), it will create a single annotated plot. For \code{verbosePlot=TRUE},
    it will create three plots (using \code{par(mfrow=c(1,3))}):
    \enumerate{
        \item raw phase counts with de-trending and de-noising loess fits
        \item residual phase counts with de-noising loess fit and detected peaks
        \item linear fit to peaks and NRL estimation
    }
}
\seealso{
\code{\link{calcPhasogram}} to calculate the phasogram from alignments,
    \code{\link{estimateNRL}} to estimate nucleosome repeat length
}
\author{
Michael Stadler
}
