% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selVarGenes.R
\name{plotSelVarGenesGroups}
\alias{plotSelVarGenesGroups}
\title{Plot Groups Genes Have Been Assigned to}
\usage{
plotSelVarGenesGroups(
  selVarGenes_list = NULL,
  xlab = "logMean",
  ylab = "logCV",
  main = "Group Assignment per Gene",
  asp = 1,
  ...
)
}
\arguments{
\item{selVarGenes_list}{the output list from the \code{selVarGenes} function.}

\item{xlab}{label for x-axis.}

\item{ylab}{label for y-axis.}

\item{main}{title for plot.}

\item{asp}{the y/x aspect ratio.}

\item{...}{additional parameters for the \code{plot} function.}
}
\value{
plot
}
\description{
This function take the output from \code{selVarGenes} and plots the group or bin each
  gene has been assigned to on the scatter plot showing the log2 coefficient of variation as a
  function of the log2 mean.
}
\details{

}
\examples{

}
\author{
Dania Machlab
}
