% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGeneRegion.R
\name{prepareGTF}
\alias{prepareGTF}
\title{Prepare GTF file for use with plotGeneRegion}
\usage{
prepareGTF(gtf)
}
\arguments{
\item{gtf}{Character scalar, path to gtf file (tested with Ensembl/Gencode files)}
}
\description{
Prepare GTF file for use with plotGeneRegion
}
\author{
Charlotte Soneson
}
