% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSampleTsvs.R
\name{readSampleTsvs}
\alias{readSampleTsvs}
\title{Read sample tsv files from seqdata storage}
\usage{
readSampleTsvs(seqdataDir = "/work/gbioinfo/seqdata", sampleIds,
  keepMulti = FALSE, ...)
}
\arguments{
\item{seqdataDir}{Character scalar, the path to the directory containing the
tsv files.}

\item{sampleIds}{Character vector with sample IDs, which will be matched
against the file names in `seqDataDir`. The sample IDs should not contain
the `.tsv` suffix.}

\item{keepMulti}{Logical scalar, indicating whether to keep samples that
match more than one tsv file. If `TRUE`, these samples are represented by
multiple rows in the table. If `FALSE`, these samples are excluded. In any
case, a warning will be generated, listing the samples with multiple
matching files.}

\item{...}{Additional arguments that will be passed to `list.files`, e.g. to
make the search case-insensitive or search recursively.}
}
\value{
A `data.frame` with metadata for the provided `sampleIds`.
}
\description{
The function searches the provided `seqdataDir` for tsv files
  corresponding to the provided `sampleIds` and returns a `data.frame`
  containing the metadata for all samples.
}
\examples{
\dontrun{
readSampleTsvs(seqdataDir = "/work/gbioinfo/seqdata", 
               sampleIds = c("1034F6-2_161111_D00404_0278_AH2F5CBCXY_ACGGACTT"))
}

}
\author{
Charlotte Soneson
}
