% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setCondaLdLibraryPath.R
\name{setCondaLdLibraryPath}
\alias{setCondaLdLibraryPath}
\title{Set LD_LIBRARY_PATH in conda environment}
\usage{
setCondaLdLibraryPath(envPath, condaLibPath = NULL)
}
\arguments{
\item{envPath}{Path to a conda environment for which the 
\code{LD_LIBRARY_PATH} variable will be set.}

\item{condaLibPath}{Path to the conda library folder that will be prepended 
to the \code{LD_LIBRARY_PATH} environment variable. If \code{NULL} (the 
default), it will be set to the subdirectory \code{lib/} of the 
\code{basilisk}-managed conda instance, which is suitable for setting 
the \code{LD_LIBRARY_PATH} for \code{basilisk} packages.
For other applications, please set the conda path accordingly. The 
path should point to a folder containing the C++ standard libraries.}
}
\value{
Nothing is returned. The function is called for its side effect 
    of creating shell scripts in the provided conda environment.
}
\description{
Create two shell scripts in subfolders of a given conda environment, 
setting and unsetting the LD_LIBRARY_PATH environment variable when 
activating and deactivating the environment.
}
\details{
See https://conda.io/projects/conda/en/latest/user-guide/tasks/manage-environments.html#saving-environment-variables
for more information about setting environment variables in conda 
environments. To get the environment path for a \code{basilisk} package, 
use e.g. \code{basilisk::obtainEnvironmentPath()}.
}
\author{
Charlotte Soneson
}
