% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissue_specificity_score.R
\docType{methods}
\name{specificityScore}
\alias{specificityScore}
\alias{specificityScore,matrix-method}
\alias{specificityScore,SummarizedExperiment-method}
\alias{specificityScore,SingleCellExperiment-method}
\title{Calculate gene-expression specificity scores.}
\usage{
specificityScore(
  x,
  method = c("tau", "TSI", "counts"),
  group = NULL,
  thresh = 0,
  expr_values = "logcounts",
  na.rm = FALSE
)

\S4method{specificityScore}{matrix}(
  x,
  method = c("tau", "TSI", "counts"),
  group = NULL,
  thresh = 0,
  expr_values = "logcounts",
  na.rm = FALSE
)

\S4method{specificityScore}{SummarizedExperiment}(
  x,
  method = c("tau", "TSI", "counts"),
  group = NULL,
  thresh = 0,
  expr_values = "logcounts",
  na.rm = FALSE
)

\S4method{specificityScore}{SingleCellExperiment}(
  x,
  method = c("tau", "TSI", "counts"),
  group = NULL,
  thresh = 0,
  expr_values = "logcounts",
  na.rm = FALSE
)
}
\arguments{
\item{x}{Expression data, either a \code{matrix} with expression values for
genes (rows) in each sample (columns), or a \code{SummarizedExperiment} or
\code{SingleCellExperiment} object containing such expression data in one
of the assays (selected by \code{expr_values}).}

\item{method}{\code{character} scalar selecting the type of expression
specificity score to be calculated. One of: "tau", "TSI", "counts".
See "Details" for method-specific information.}

\item{group}{\code{character} or \code{factor} of length \code{ncol(x)} that
groups the measurements into clusters or tissues, for which expression
specificity scores are to be calculated. If \code{NULL} (the default), each
column of \code{x} is assumed to be its own group. If multiple columns belong
to the same group, these columns are first aggregated (averaged) before
score calculations.}

\item{thresh}{\code{numeric} scalar defining the expression threshold. Values
greater than this threshold are interpreted as expressed (used only for
some of the methods, see "Details").}

\item{expr_values}{Integer scalar or string indicating which assay of
\code{x} contains the expression values, for \code{x} of type
\code{SummarizedExperiment} or \code{SingleCellExperiment}. Ignored if
\code{x} is a \code{matrix}.}

\item{na.rm}{Logical scalar. If \code{TRUE}, \code{NA} values are excluded in
the calculations.}
}
\value{
A \code{numeric} vector of length \code{nrow(x)} with gene scores.
}
\description{
Calculate expression specificity scores for genes that quantify
   specific expression of a gene in groups of samples (e.g. from different tissues).
}
\examples{
x <- rbind(g1 = runif(5),
           g2 = c(1, 0, 0, 0, 0),
           g3 = c(.6, .1, .1, .1, .1))
specificityScore(x)
specificityScore(x, method = "TSI")
specificityScore(x, method = "counts", thresh = 0.5)

}
\references{
For a review of tissue-specificity scores, see:
"A benchmark of gene expression tissue-specificity metrics"
Nadezda Kryuchkova-Mostacci and Marc Robinson-Rechavi
Brief Bioinform. 2017 Mar; 18(2): 205–214.
doi: 10.1093/bib/bbw008, PMCID: PMC5444245, PMID: 26891983
}
\author{
Michael Stadler
}
