% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissknife.R
\docType{package}
\name{swissknife-package}
\alias{swissknife}
\alias{swissknife-package}
\title{swissknife - handy code shared in the FMI CompBio group}
\description{
\pkg{swissknife} is a collection of useful R functions performing various
    tasks that might be re-usable and worth sharing.
}
\author{
\strong{Maintainer}: Michael Stadler \email{michael.stadler@fmi.ch}

Authors:
\itemize{
  \item Charlotte Soneson \email{charlotte.soneson@fmi.ch}
  \item Panagiotis Papasaikas \email{panagiotis.papasaikas@fmi.ch}
  \item Dania Machlab \email{dania.machlab@fmi.ch}
}

}
