% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addUtilsFunctions.R
\name{addUtilsFunctions}
\alias{addUtilsFunctions}
\title{Copy utility functions to package}
\usage{
addUtilsFunctions(outfile = "R/utils.R")
}
\arguments{
\item{outfile}{Character scalar, giving the path to which the script 
should be copied. The path is relative to the root of the active 
project. If a file with this name already exists, the function will 
ask for confirmation before overwriting it.}
}
\description{
This function copies handy utility functions to a new script in a 
specified location. Currently, the script contains the following utility 
functions:
\itemize{
\item \code{.assertScalar()} - convenience function to check the validity 
    of scalar variables.
\item \code{.assertVector()} - convenience function to check the validity 
    of vector variables. 
}
}
\author{
Charlotte Soneson
}
