% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calcAndCountDist}
\alias{calcAndCountDist}
\title{[internal] Count frequency of differences between values in integer vectors.}
\usage{
calcAndCountDist(x, p, cnt)
}
\arguments{
\item{x}{first \code{integer} vector.}

\item{p}{second \code{integer} vector. Distances are calculated from each element
in \code{p} to each greater element in \code{x}.}

\item{cnt}{\code{NumericVector} to store the result in. The length of \code{cnt} defines
the maximal distance that will be included in the analysis, and new counts
will be added to the values of \code{cnt}.}
}
\value{
\code{numeric} vector \code{cnt}, where \code{cnt[d]} correspond to
    the number of observed distances \code{d}.
}
\description{
Given two ascendingly sorted integer vectors \code{x} and \code{p}, calculate
    and cound the differences between their elements that are greater than zero
    and less than \code{maxd}. The number of observed distances \code{d} are reported in \code{cnt[d]},
    and `maxd` corresponds to the \code{length(cnt)}.
}
\author{
Michael Stadler
}
