% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phasograms.R
\name{estimateNRL}
\alias{estimateNRL}
\title{Estimate the nucleosome repeat length (NRL) from a phasogram.}
\usage{
estimateNRL(x, mind = 140L, usePeaks = 1:8, span1 = 100/length(x),
  span2 = 1500/length(x))
}
\arguments{
\item{x}{\code{numeric} vector giving the counts of alignment distances
(typically the output of \code{\link{calcPhasogram}}.}

\item{mind}{\code{integer(1)} specifying the minimal distance to be used for
NRL estimation. The default value (140) ignores any distance too short to
span at least a single nucleosome.}

\item{usePeaks}{\code{integer} vector selecting the modes (peaks) in the
phasogram used in NRL estimation.}

\item{span1}{\code{numeric(1)} giving the smoothing parameter for de-trending
loess fit (high pass filter).}

\item{span2}{\code{numeric(1)} giving the smoothing parameter for de-noising
loess fit (low pass filter).}
}
\value{
`list` with elements:
    - `nrl` the estimated nucleosome repeat length
    - `nrl.CI95` the 95% confidence interval
    - `xs` smoothed (de-trended) phasogram
    - `loessfit` the de-noising fit to the de-trended phasogram
    - `lmfit` the linear fit to the phasogram peaks
    - `peaks` the peak locations
    - `mind` minimal distance included in the fit
    - `span1` smoothing parameter for de-trending loess fit
    - `span2` smoothing parameter for de-noising loess fit
    - `usePeaks` the peaks used in the fit
}
\description{
Estimate the nucleosome repeat length (NRL) from the frequencies
  of same-strand alignment distances (phasogram), e.g. generated by
  \code{\link{calcPhasogram}}. The NRL is obtained from the slope of a linear
  fit to the modes in the phasogram.
}
\seealso{
\code{\link{calcPhasogram}} to calculate the phasogram from
  alignments, \code{\link{plotPhasogram}} to visualize an annotated phasogram
}
\author{
Michael Stadler
}
