% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_labeling.R
\name{labelCells}
\alias{labelCells}
\title{Assign labels to cells using known marker genes}
\usage{
labelCells(
  sce,
  markergenes,
  fraction_topscoring = 0.01,
  expr_values = "logcounts",
  normGenesetExpressionParams = list(R = 200),
  aggregateReferenceParams = list(power = 0.5),
  SingleRParams = list(),
  BPPARAM = SerialParam()
)
}
\arguments{
\item{sce}{\code{SingleCellExperiment} object.}

\item{markergenes}{Named \code{list} of \code{character} vectors with the
marker genes for each cell types. The marker genes must be a subset of
\code{rownames(sce)}.}

\item{fraction_topscoring}{\code{numeric} vector of length 1 or the same
length as \code{markergenes} giving the fraction(s) of top scoring cells
for each cell type to pick to create the reference transcriptome profile.}

\item{expr_values}{Integer scalar or string indicating which assay of
\code{sce} contains the expression values.}

\item{normGenesetExpressionParams}{\code{list} with additional parameters
for \code{\link{normGenesetExpression}}.}

\item{aggregateReferenceParams}{\code{list} with additional parameters
for \code{\link[SinlgeR]{aggregateReference}}.}

\item{SingleRParams}{\code{list} with additional parameters for
\code{\link[SingleR]{SingleR}}.}

\item{BPPARAM}{An optional \code{\link[BiocParallel]{BiocParallelParam}}
instance determining the parallel back-end to be used during evaluation.}
}
\value{
A \code{list} of three elements named \code{cells}, \code{refs} and \code{labels}.
  \code{cells} contains a \code{list} with the numerical indices of the top
  scoring cells for each cell type.
  \code{refs} contains the pseudo-bulk transcriptome profiles used as a 
  reference for label assignment, as returned by \code{\link[SinlgeR]{aggregateReference}}.
  \code{labels} contains a \code{\link[S4Vectors]{DataFrame}} with the
  annotation statistics for each cell (one cell per row), generated by
  \code{\link[SingleR]{SingleR}}.
}
\description{
Given marker gene sets for cell types, identify cells with
  high expression of the marker genes (positive examples), then use these
  cells to create a reference transcriptome profile for each cell type and
  identify additional cells of each type using \code{SingleR}.
}
\seealso{
\code{\link{normGenesetExpression}} used to calculate scores for
   marker gene sets; \code{\link[SinlgeR]{aggregateReference}} used to
   create reference profiles; \code{\link[SingleR]{SingleR}} used to assign
   labels to cells.
}
\author{
Michael Stadler
}
