% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBitScatter.R
\name{plotBitScatter}
\alias{plotBitScatter}
\title{Create a bitmap-rendered plot.}
\usage{
plotBitScatter(x, y = NULL, ..., densCols = TRUE,
  colpal = c("#00007F", "blue", "#007FFF", "cyan", "#7FFF7F", "yellow",
  "#FF7F00", "red", "#7F0000"), xpixels = 1000, ypixels = NULL)
}
\arguments{
\item{x}{\code{numeric} vector with x-coordinates of points, or a two-column
matrix with x- and y- coordinates.}

\item{y}{\code{numeric} vector with y-coordinates of points (same length as \code{x}).
Can be \code{NULL}, in which case \code{x} must be a two-column matrix.}

\item{...}{any further arguments to be passed to \code{plot}}

\item{densCols}{\code{logical(1)}. If \code{TRUE} and \code{col} is not given
as an additional argument, then the local density of points will be used
as colors, using the palette spanned by the colors in \code{colpal}.}

\item{colpal}{vector of colors defining the palette for automatic density-based coloring.}

\item{xpixels}{the number of pixels in the x dimension used for rendering
the plotting area.}

\item{ypixels}{the number of pixels in the y dimension used for rendering the
plotting area. If \code{NULL} (the default), will be calculated automatically
as \code{xpixels * par('pin')[2] / par('pin')[1]}, such that the aspect ratio of
the current plotting region is observed. This may not work (e.g. when using
\code{layout()}), as this may result in negative values returned by \code{par('pin')}.
In that case, \code{ypixels} should be set manually using this argument.}
}
\value{
\code{NULL} (invisibly)
}
\description{
\code{plotBitScatter} is a wrapper around \code{plot} which renders the
    plot area as a bitmap (png), but keeps all other elements (axes, labels, etc.)
    as vector elements. This is especially useful for keeping the size of PDF files
    with scatter plots with many elements small, while retaining editability of axes.
}
\details{
\code{xpixels} controls the resolution of the rendered plotting area.
    In order to keep circular plotting symbols circlular (e.g. \code{pch = 1}),
    \code{ypixels} is automatically calculated using \code{xpixels} and the
    aspect ratio of the current plotting area. If the plotting device is
    rescaled after calling \code{plotBitScatter}, circular plotting symbols
    may become skewed.
}
\examples{
x <- rnorm(1000)
y <- rnorm(1000)
par(mfrow=c(1,2))
plotBitScatter(x, y, main = "bitmap")
plot(x, y, main = "default")

}
\author{
Michael Stadler
}
