% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selVarGenes.R
\name{plotSelVarGenes}
\alias{plotSelVarGenes}
\title{Plot Selected Variable Genes}
\usage{
plotSelVarGenes(
  selVarGenes_list = NULL,
  xlab = "logMean",
  ylab = "logCV",
  main = "Selected Variable Genes",
  pch = 16,
  col = "#BEBEBE40",
  sel_col = "steelblue",
  colByBin = FALSE,
  asp = 1,
  ...
)
}
\arguments{
\item{selVarGenes_list}{the output list from the \code{selVarGenes} function.}

\item{xlab}{label for x-axis.}

\item{ylab}{label for y-axis.}

\item{main}{title for plot.}

\item{pch}{point pch.}

\item{col}{point color.}

\item{sel_col}{point color of the selected variable genes.}

\item{colByBin}{if TRUE, color the genes by the bin they've been assigned to.}

\item{asp}{the y/x aspect ratio. Set to 1 when \code{colByBin} is TRUE.}

\item{...}{additional parameters for the \code{plot} function.}
}
\value{
plot
}
\description{
This function take the output from \code{selVarGenes} and plots the genes that have been 
  selected to be highly variable across the cells. It plot the log2 coefficient of variation as
  a function of the log mean.
}
\examples{
if (requireNamespace("SingleCellExperiment", quietly = TRUE)) {
    # packages
    library(SingleCellExperiment)
   
    # create example count matrix
    # ... poisson distr per gene
    mu <- ceiling(runif(n = 2000, min = 0, max = 100))
    counts <- do.call(rbind, lapply(mu, function(x){rpois(1000, lambda = x)}))
    counts <- counts + 1
    # ... add signal to subset of genes (rows) and cells (columns)
    i <- sample(x = 1:nrow(counts), size = 500)
    j <- sample(x = 1:ncol(counts), size = 500)
    counts[i, j] <- counts[i, j] + sample(5:10, length(i), replace = TRUE)
 
    # create SCE
    sce <- SingleCellExperiment(list(counts = counts))

    # calculate sizeFactors
    libsizes <- colSums(counts)
    sizeFactors(sce) <- libsizes / mean(libsizes)

    # select variable genes
    varGenes <- selVarGenes(sce)
   
    # plot
    plotSelVarGenes(varGenes)
    plotSelVarGenes(varGenes, colByBin=TRUE)
}

}
\author{
Dania Machlab
}
