% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selVarGenes.R
\name{plotSelVarGenesGroups}
\alias{plotSelVarGenesGroups}
\title{Plot Groups Genes Have Been Assigned to}
\usage{
plotSelVarGenesGroups(
  selVarGenes_list = NULL,
  xlab = "logMean",
  ylab = "logCV",
  main = "Group Assignment per Gene",
  asp = 1,
  ...
)
}
\arguments{
\item{selVarGenes_list}{the output list from the \code{selVarGenes} function.}

\item{xlab}{label for x-axis.}

\item{ylab}{label for y-axis.}

\item{main}{title for plot.}

\item{asp}{the y/x aspect ratio.}

\item{...}{additional parameters for the \code{plot} function.}
}
\value{
plot
}
\description{
This function take the output from \code{selVarGenes} and plots the group or bin each
  gene has been assigned to on the scatter plot showing the log2 coefficient of variation as a
  function of the log2 mean.
}
\examples{
   # packages
   library(SingleCellExperiment)
   
   # create example count matrix
   # ... poisson distr per gene
   mu <- ceiling(runif(n = 2000, min = 0, max = 100))
   counts <- do.call(rbind, lapply(mu, function(x){rpois(1000, lambda = x)}))
   counts <- counts + 1
   # ... add signal to subset of genes (rows) and cells (columns)
   i <- sample(x = 1:nrow(counts), size = 500)
   j <- sample(x = 1:ncol(counts), size = 500)
   counts[i, j] <- counts[i, j] + sample(5:10, length(i), replace = TRUE)
   
   # create SCE
   sce <- SingleCellExperiment(list(counts=counts))
   
   # calculate sizeFactors
   libsizes <- colSums(counts)
   sizeFactors(sce) <- libsizes/mean(libsizes)
 
   # select variable genes
   varGenes <- selVarGenes(sce)
   
   # plot
   plotSelVarGenesGroups(varGenes)

}
\author{
Dania Machlab
}
