% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleControlElements.R
\name{sampleControlElements}
\alias{sampleControlElements}
\title{Sample control elements that match a target distribution.}
\usage{
sampleControlElements(
  x,
  idxTarget,
  idxControl = NULL,
  nbins = 50,
  oversample = 1
)
}
\arguments{
\item{x}{\code{numeric} vector (or list of \code{numeric} vectors).
\code{idxTarget} and \code{idxControl} refer to the elements of \code{x}.
If \code{x} is a list, all elements must have the same length.}

\item{idxTarget}{\code{numeric} or \code{logical} vector specifying the elements
in \code{x} that define the target distribution to be matched by the
control elements.}

\item{idxControl}{\code{numeric} or \code{logical} vector specifying the
complete set of possible control elements in \code{x} (default: all that
are not in \code{idxTarget}), from which a subset is to be sampled.}

\item{nbins}{\code{numeric(1)} or \code{numeric(length(x))} if \code{x} is a
list, specifying the number of bins to group the values of x into. Higher
 numbers of bins will increase the match to the target distribution(s),
 but may fail if there are few elements to sample from (will throw a warning).}

\item{oversample}{The number of control elements to sample for each target element.}
}
\value{
\code{numeric} vector with \code{round(length(idxTarget) * oversample)}
    elements, specifying the index (positions) of the sampled control elements.
}
\description{
Randomly sample from a set of control (background) elements,
    such that the selected elements are similarly distributed as a given
    set of target (foreground) elements.
}
\examples{
x <- c(runif(1000, min = 0, max = 10),
       rnorm(200, mean = 5, sd = 1))
s <- sampleControlElements(x, idxTarget = 1001:1200, idxControl = 1:1000)
par(mfrow=c(2,2))
h <- hist(x, breaks = 20, main = "all")
hist(x[1:1000], breaks = h$breaks, main = "all control")
hist(x[1001:1200], breaks = h$breaks, main = "target")
hist(x[s], breaks = h$breaks, main = "sampled control")

}
\author{
Michael Stadler
}
