% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selVarGenes.R
\name{selVarGenes}
\alias{selVarGenes}
\title{Select Variable Genes in Single Cell RNA-seq}
\usage{
selVarGenes(
  data = NULL,
  assay.type = "counts",
  Nmads = 3,
  minCells = 5,
  minExpr = 1,
  exclTopExprFrac = 0.01,
  span = 0.2,
  control = stats::loess.control(surface = "direct"),
  nBins = 100,
  nBinsDense = ceiling(nrow(data)/4),
  ...
)
}
\arguments{
\item{data}{\code{SingleCellExperiment} object or (normalized) count \code{matrix} containing the genes as rows and cells as columns.}

\item{assay.type}{the type of assay to use if \code{data} is a \code{SingleCellExperiment}. It can be either 'counts' 
or 'logcounts'. The default is 'counts'.}

\item{Nmads}{number of MADs beyond which genes are selected per bin.}

\item{minCells}{keep genes with minimum expression in at least this number of cells.}

\item{minExpr}{keep genes with expression greater than or equal to this in \code{minCells} cells in the 
normalized count matrix.}

\item{exclTopExprFrac}{the fraction of top expressed genes that will be excluded from the loess fit 
(value between 0 and 1).}

\item{span}{span parameter for \code{loess} function.}

\item{control}{control parameters for \code{loess} function.}

\item{nBins}{number of bins or groups to place the points(genes) into.}

\item{nBinsDense}{number of bins or groups to use  to place the points(genes) into
when calculating more accurate distance values to the curve from the loess fit.}

\item{...}{additional parameters for the \code{loess} function from the \code{stats} package.}
}
\value{
a \code{list} of length 2:
  \itemize{
     \item varGenes: \code{vector} containing the names of the most variable genes.
     \item geneInfo: \code{data.frame} with genes as rows and columns containing  
                     calculated measures:
                     \itemize{
                        \item logMean: log2(mean) expression of genes across cells.
                        \item logCV: log2(coefficient of variation) of genes across cells.
                        \item pred_logCV: predicted log2(coefficient of variation) from loess fit.
                        \item assigned_bin: bin each gene has been assigned to.
                        \item distance: accurate distance measuses. Points below the loess fit get a negative sign.
                     }
  }
}
\description{
This function selects the most variable genes from a \code{SingleCellExperiment} object
  using the plot that displays the log2 coefficient of variation as a function of the log2 mean for 
  all genes across all the cells.
}
\details{
The function takes in a \code{SingleCellExperiment} object and calculates the normalized
  counts by dividing the raw counts by the corresponding sizeFactors per cell. Only genes that have
  an expression greater than or equal to \code{minExpr} in at least \code{minCells} cells will be
  kept.
  
  The genes that vary most on the log2(coefficient of variation) vs log2(mean) plot of genes will 
  be selected. A loess fit is done on this plot and the distance (euclidean by default) each point has
  to the curve is calculated in two steps.
  
  In the first step, genes are assigned to bins by taking the minimum distance to the curve. By default
  we select 100 points on the loess fit and calculate the distances each gene has to all those points on 
  the curve. Each gene is assigned to the point on the curve for which it has the shortest distance. In 
  the second step, more accurate distances to the curve are calculated by using a higher number of 
  points on the curve. Distances are calculated using the \code{\link[wordspace]{dist.matrix}} function.
  
  Finally, for each bin, the most variable genes are selected using the more accurate distance
  measures. Genes that fall below the loess fit are assigned a negative sign and the genes that 
  are \code{Nmads} MADs away from the median are selected.
}
\examples{
   # packages
   library(SingleCellExperiment)
   
   # create example count matrix
   # ... poisson distr per gene
   mu <- ceiling(runif(n = 2000, min = 0, max = 100))
   counts <- do.call(rbind, lapply(mu, function(x){rpois(1000, lambda = x)}))
   counts <- counts + 1
   # ... add signal to subset of genes (rows) and cells (columns)
   i <- sample(x = 1:nrow(counts), size = 500)
   j <- sample(x = 1:ncol(counts), size = 500)
   counts[i, j] <- counts[i, j] + sample(5:10, length(i), replace = TRUE)
   
   # create SCE 
   sce <- SingleCellExperiment(list(counts=counts))
   
   # calculate sizeFactors
   libsizes <- colSums(counts)
   sizeFactors(sce) <- libsizes/mean(libsizes)
 
   # select variable genes
   varGenes <- selVarGenes(sce, assay.type="counts")
   
   # plot
   plotSelVarGenes(varGenes, colByBin=TRUE)
   plotSelVarGenes(varGenes)

}
\author{
Dania Machlab
}
