% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightedMeanByID.R
\name{weightedMeanByID}
\alias{weightedMeanByID}
\title{Aggregate different rows assigned to the same ID by calculating a 
weighted mean}
\usage{
weightedMeanByID(
  SE,
  assay,
  idCol = "GENEID",
  log2Transformed = TRUE,
  pseudocount = 0.01
)
}
\arguments{
\item{SE}{\code{SummarizedExperiment} object that contains an assay with 
the values to be aggregated and a rowData column with IDs to indicate
which values of different rows to combine.}

\item{assay}{\code{character} name of the assay in the SummarizedExperiment
object to use.}

\item{idCol}{\code{character} column name in the rowData of the
SummarizedExperiment indicating the ID assigned to each row.}

\item{log2Transformed}{\code{logical} indicating whether the data is log2
transformed.}

\item{pseudocount}{\code{numeric} pseudocount to add to avoid dividing by 0.}
}
\value{
\code{data.frame} with one column that is a numeric vector of the
weighted means per unique ID. If log2Transformed is true the output will be 
log2 transformed again. The rownames of the dataframe correspond to the 
unique IDs.
}
\description{
After calculating row means to summarize across replicates, 
values of different rows that are assigned to the same ID are summarized by 
calculating a weighted mean. The weights are the values themselves.
}
\author{
Fiona Ross
}
